move msg conversion to seperate file
